divide highlight + added highlight property + improve mouseover  function for  work with Chrome
