+++ src/test/resources/application.properties
@@ -3,3 +3,4 @@ site.url=${site.url}
 grid.url=${grid.url}
 timeToTimeout=30
 timeToSleep=0
+highlight=${highlight}
